+++ examples/misc/compare_swmmio.py
@@ -1,0 +1,3 @@
+from swmmio import Model
+
+m = Model('epaswmm5_apps_manual/Example7-Final.inp', include_rpt=True)