+++ examples/read_example_controls.py
@@ -1,0 +1,5 @@
+from swmm_api import SwmmInput
+
+inp = SwmmInput.read_file('example_controls.inp')
+c = inp.CONTROLS
+print(c.to_inp_lines())